backend/requirements.txt
# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.4.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Redis
redis>=5.0.8

# HTTP Client
httpx>=0.26.0

# Task Queue
celery>=5.3.0

# Google AI
google-generativeai>=0.8.0
langchain>=0.2.14

# Sarvam AI (OpenAI-compatible client)
openai>=1.30.0

# QR Code
qrcode[pil]>=7.4.2

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Logging
structlog>=24.1.0
